+++ src/main/java/com/zcpu/tzzmod/command/TzzRootCommand.java
@@ -3,6 +3,7 @@ package com.zcpu.tzzmod.command;
 import com.mojang.brigadier.CommandDispatcher;
 import com.zcpu.tzzmod.region.RegionControllerCommand;
 import com.zcpu.tzzmod.signal.SignalCommand;
+import com.zcpu.tzzmod.webadmin.WebAdminCommand;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 
@@ -17,6 +18,7 @@ public final class TzzRootCommand {
                 .then(NoteCommand.build())
                 .then(SendMsgCommand.build())
                 .then(RegionControllerCommand.build())
-                .then(SignalCommand.build()));
+                .then(SignalCommand.build())
+                .then(WebAdminCommand.build()));
     }
 }